+++ src/query/storages/stage/src/input_context_bridge.rs
