+++ src/query/storages/stage/src/lib.rs
@@ -16,7 +16,7 @@
 #![allow(clippy::uninlined_format_args)]
 
 mod append;
-mod input_context;
+mod input_context_bridge;
 mod one_file_partition;
 mod read;
 mod stage_table;
